Disable libostree docs
The API docs are LGPL2 which, if we're being strict, would make the
entire result LGPL2 if the docs are included in the binary; I assume, at
least gtk-rs makes a point to not include the docs in the main build. It
should be possible to make a build script and associated feature that
includes the API docs just for the docs build.