ostree/main: Prefix version with "v"
Prefix the `Version` key in the YAML-compatible output of
`ostree --version` with `v` so that it's coerced to a string. The issue
with the previous approach in a nutshell:

```
In [5]: yaml.load("asdf: 2018.10")
Out[5]: {'asdf': 2018.1}
```

It's treating the version number as a floating-point. Now, this is
technically a backwards incompatible change, but given that the previous
approach is inherently broken for our needs, I don't see a way around
breaking it now.