ostree commit: support --bind-collection-ref=COLLECTION-ID:BRANCH
This allows you to create a commit bound to whatever collection id/ref
it would otherwise bind to, plus a set of extra pair.

Only one of --bind-collection-ref and --bind-ref is supported, but if
you do the former we also set the regular ref bindings based on it.

We also try to set the old ostree.collection-binding key to a useful value.