pull: When pulling a collection-id + ref, verify bindings against that
I.e. when you pass 'collection-refs' instead of 'refs' to
ostree_repo_pull_with_options to pull a
refs/mirrors/$collection_id/$ref kind of ref, we need to verify
against that collection id, not whatever happened to be the default
collection id for that remote (or the repo-wide collection id for a
local repo).