Retry requests on HTTP 503 and 504
PR #1594 added logic to retry downloads when they fail due to network
errors, or due to HTTP 408 Request Timeout.

In practice, pulling from Flathub frequently fails with 503 Service
Unavailable. This is a transient error, so it makes sense to retry when
it's encountered.

There is no good code in GIOErrorEnum for this so this patch takes the
lazy route of mapping it to G_IO_ERROR_TIMED_OUT, which is already
treated as a transient error in _ostree_fetcher_should_retry_request().

While we're here, also map 504 Gateway Timeout to G_IO_ERROR_TIMED_OUT
(which does seem correct, if lossy).