lib/gpg: Allow local server override for WKD URLs
In order to test `ostree_remote_update_gpg_keys`, we need to be able to
fetch the keys from a local test server. This inherently requires
introducing a backdoor to the update process. If the
_OSTREE_GPG_UPDATE_LOCAL_PORT environment variable is set, change the
server to http://127.0.0.1:<port> after validating that the port is
numerical. This should keep any attack local by not allowing the URL to
be changed to an arbitrary remote server.